Add control check version
general_opts and prod_opt_list now comes with a "MODIStspVersion" attribute: if the version of general options is older than the current MODIStsp, a warning is generated; if it is older than 1.2.2, MODIStsp is stopped. If the version of prod_opt_list is older than current MODIStsp, it is re-generated.